+++ TinkoffASDKCore/TinkoffASDKCore/AcquiringSdk.swift
@@ -22,27 +22,25 @@ import UIKit
 
 public enum AcquiringSdkError: Error {
     case publicKey(String)
-
     case url
 }
 
-///
 /// `AcquiringSdk`  позволяет конфигурировать SDK и осуществлять взаимодействие с **Тинькофф Эквайринг API**  https://oplata.tinkoff.ru/landing/develop/
 public final class AcquiringSdk: NSObject {
     public var fpsEnabled: Bool = false
-
-    private var networkTransport: NetworkTransport
-    private var terminalKey: String
-    private var publicKey: SecKey
-    public private(set) var languageKey: AcquiringSdkLanguage?
-    private var logger: LoggerDelegate?
+    public var fpsEnabled: Bool = false
+    public let languageKey: AcquiringSdkLanguage?
+    private let networkTransport: NetworkTransport
+    private let terminalKey: String
+    private let publicKey: SecKey
     private let tokenGenerator: TokenGenerator?
 
     /// Создает новый экземпляр SDK
     public init(configuration: AcquiringSdkConfiguration) throws {
-        fpsEnabled = configuration.fpsEnabled
-
-        terminalKey = configuration.credential.terminalKey
+        self.fpsEnabled = configuration.fpsEnabled
+        self.terminalKey = configuration.credential.terminalKey
+        self.languageKey = configuration.language
+        self.tokenGenerator = configuration.tokenGenerator
 
         if let publicKey: SecKey = RSAEncryption.secKey(string: configuration.credential.publicKey) {
             self.publicKey = publicKey
@@ -52,26 +50,26 @@ public final class AcquiringSdk: NSObject {
 
         if let url = URL(string: "https://\(configuration.serverEnvironment.rawValue)/"),
            let certsConfigUrl = URL(string: "https://\(configuration.configEnvironment.rawValue)/") {
-            let deviceInfo = DeviceInfo(model: UIDevice.current.localizedModel,
-                                        systemName: UIDevice.current.systemName,
-                                        systemVersion: UIDevice.current.systemVersion)
+            let deviceInfo = DeviceInfo(
+                model: UIDevice.current.localizedModel,
+                systemName: UIDevice.current.systemName,
+                systemVersion: UIDevice.current.systemVersion
+            )
 
             let sessionConfiguration = URLSessionConfiguration.default
             sessionConfiguration.timeoutIntervalForRequest = configuration.requestsTimeoutInterval
             sessionConfiguration.timeoutIntervalForResource = configuration.requestsTimeoutInterval
             
-            networkTransport = AcquaringNetworkTransport(urlDomain: url,
-                                                         certsConfigDomain: certsConfigUrl,
-                                                         session: URLSession(configuration: sessionConfiguration),
-                                                         deviceInfo: deviceInfo)
+            networkTransport = AcquaringNetworkTransport(
+                urlDomain: url,
+                certsConfigDomain: certsConfigUrl,
+                session: URLSession(configuration: sessionConfiguration),
+                deviceInfo: deviceInfo,
+                logger: configuration.logger
+            )
         } else {
             throw AcquiringSdkError.url
         }
-
-        languageKey = configuration.language
-        logger = configuration.logger
-        networkTransport.logger = logger
-        tokenGenerator = configuration.tokenGenerator
     }
 
     private func createParameters(with request: AcquiringRequestTokenParams, completion: @escaping (JSONObject) -> Void) {
@@ -115,7 +113,7 @@ public final class AcquiringSdk: NSObject {
         return task
     }
 
-    /// Обновляем информцию о реквизитах карты, добавляем шифрование
+    /// Обновляем информацию о реквизитах карты, добавляем шифрование
     private func updateCardDataRequestParams(_ parameters: inout JSONObject?) {
         if let cardData = parameters?[PaymentFinishRequestData.CodingKeys.cardData.rawValue] as? String {
             if let encodedCardData = RSAEncryption.encrypt(string: cardData, publicKey: publicKey) {
@@ -124,7 +122,7 @@ public final class AcquiringSdk: NSObject {
         }
     }
 
-    /// Получить IP адресс
+    /// Получить IP адрес
     public func networkIpAddress() -> String? {
         return networkTransport.myIpAddress()
     }
@@ -137,37 +135,50 @@ public final class AcquiringSdk: NSObject {
     ///   - data: `PaymentInitPaymentData` информация о заказе на оплату
     ///   - completionHandler: результат операции `PaymentInitResponse` в случае удачной регистрации и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func paymentInit(data: PaymentInitData, completionHandler: @escaping (_ result: Result<PaymentInitResponse, Error>) -> Void) -> Cancellable {
+    public func paymentInit(
+        data: PaymentInitData,
+        completionHandler: @escaping (_ result: Result<PaymentInitResponse, Error>) -> Void
+    ) -> Cancellable {
         let paramsEnricher: IPaymentInitDataParamsEnricher = PaymentInitDataParamsEnricher()
         let enrichedData = paramsEnricher.enrich(data)
         
         let request = PaymentInitRequest(data: enrichedData)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completionHandler)
+>>>>>>> dev
     }
 
     // MARK: - подтверждение платежа
 
-    /// Создать запрос для подтвержения платежа 3ds формы
+    /// Создать запрос для подтверждения платежа 3DS формы
     ///
     /// - Parameters:
     ///   - data: `Confirmation3DSData`
     /// - Returns:
     ///   - URLRequest
     public func createConfirmation3DSRequest(data: Confirmation3DSData) throws -> URLRequest {
-        return try networkTransport.createConfirmation3DSRequest(requestData: data)
+        try networkTransport.createConfirmation3DSRequest(requestData: data)
     }
 
-    /// Создать запрос для подтвержения платежа 3ds формы
+    /// Создать запрос для подтверждения платежа 3DS формы
     ///
     /// - Parameters:
     ///   - data: `Confirmation3DSData`
     /// - Returns:
     ///   - URLRequest
-    public func createConfirmation3DSRequestACS(data: Confirmation3DSDataACS, messageVersion: String) throws -> URLRequest {
-        return try networkTransport.createConfirmation3DSRequestACS(requestData: data, messageVersion: messageVersion)
+    public func createConfirmation3DSRequestACS(
+        data: Confirmation3DSDataACS,
+        messageVersion: String
+    ) throws -> URLRequest {
+        try networkTransport.createConfirmation3DSRequestACS(requestData: data, messageVersion: messageVersion)
     }
 
-    /// Проверяет параметры для 3ds формы
+    /// Проверяет параметры для 3DS формы
     ///
     /// - Parameters:
     ///   - data: `Checking3DSURLData`
@@ -177,116 +188,279 @@ public final class AcquiringSdk: NSObject {
         return try networkTransport.createChecking3DSURL(requestData: data)
     }
 
-    /// callback URL для завершения 3ds подтверждения
+    /// callback URL для завершения 3DS подтверждения
     ///
     /// - Returns:
     ///   - URL
     public func confirmation3DSTerminationURL() -> URL {
-        return networkTransport.confirmation3DSTerminationURL
+        networkTransport.confirmation3DSTerminationURL
     }
 
     public func confirmation3DSTerminationV2URL() -> URL {
-        return networkTransport.confirmation3DSTerminationV2URL
+        networkTransport.confirmation3DSTerminationV2URL
     }
 
     public func confirmation3DSCompleteV2URL() -> URL {
-        return networkTransport.complete3DSMethodV2URL
+        networkTransport.complete3DSMethodV2URL
     }
 
     /// Проверяем версию 3DS перед подтверждением инициированного платежа передачей карточных данных и идентификатора платежа
     ///
     /// - Parameters:
     ///   - data: `PaymentFinishRequestData`
-    ///   - completionHandler: результат операции `Check3dsVersionResponse` в случае удачного ответа и `Error` - в случе ошибки.
-    public func check3dsVersion(data: PaymentFinishRequestData, completionHandler: @escaping (_ result: Result<Check3dsVersionResponse, Error>) -> Void) -> Cancellable {
+    ///   - completionHandler: результат операции `Check3dsVersionResponse` в случае удачного ответа и `Error` - в случае ошибки.
+    @discardableResult
+    public func check3dsVersion(
+        data: PaymentFinishRequestData,
+        completionHandler: @escaping (_ result: Result<Check3dsVersionResponse, Error>) -> Void
+    ) -> Cancellable {
         let requestData = PaymentFinishRequestData(paymentId: data.paymentId, paymentSource: data.paymentSource)
         let request = Check3dsVersionRequest(data: requestData)
         updateCardDataRequestParams(&request.parameters)
 
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completionHandler)
+>>>>>>> dev
     }
 
     /// Подтверждает инициированный платеж передачей карточных данных
     ///
     /// - Parameters:
     ///   - data: `PaymentFinishRequestData`
-    ///   - completionHandler: результат операции `PaymentFinishResponse` в случае удачного проведеня платежа и `Error` - в случе ошибки.
-    public func paymentFinish(data: PaymentFinishRequestData, completionHandler: @escaping (_ result: Result<PaymentFinishResponse, Error>) -> Void) -> Cancellable {
+    ///   - completionHandler: результат операции `PaymentFinishResponse` в случае удачного проведения платежа и `Error` - в случае ошибки.
+    @discardableResult
+    public func paymentFinish(
+        data: PaymentFinishRequestData,
+        completionHandler: @escaping (_ result: Result<PaymentFinishResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = PaymentFinishRequest(data: data)
         updateCardDataRequestParams(&request.parameters)
 
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completionHandler)
+>>>>>>> dev
     }
 
-    ///
     /// Подтверждает инициированный платеж передачей информации о рекуррентном платеже
-    public func chargePayment(data: PaymentChargeRequestData, completionHandler: @escaping (_ result: Result<PaymentStatusResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func chargePayment(
+        data: PaymentChargeRequestData,
+        completionHandler: @escaping (_ result: Result<PaymentStatusResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = PaymentChargeRequest(data: data)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completionHandler)
+>>>>>>> dev
     }
 
     // MARK: - Статус операции
 
-    ///
     /// Получить статус платежа
-    public func paymentOperationStatus(data: PaymentInfoData, completionHandler: @escaping (_ result: Result<PaymentStatusResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func paymentOperationStatus(
+        data: PaymentInfoData,
+        completionHandler: @escaping (_ result: Result<PaymentStatusResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = PaymentStatusRequest(data: data)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completionHandler)
+>>>>>>> dev
     }
 
-    // MARK: - Cписок карт
+    // MARK: - Card List
 
-    ///
     /// - Parameters:
     ///   - data: `InitGetCardListData` информация о клиенте для получения списка сохраненных карт
-    ///   - completionHandler: результат операции `CardListResponse` в случае удачной регистрации и  `Error` - ошибка.
+    ///   - completion: результат операции `CardListResponse` в случае удачной регистрации и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func сardList(data: InitGetCardListData, responseDelegate: NetworkTransportResponseDelegate?, completionHandler: @escaping (_ result: Result<CardListResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func cardList(
+        data: InitGetCardListData,
+        responseDelegate: NetworkTransportResponseDelegate? = nil,
+        completion: @escaping (_ result: Result<CardListResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = CardListRequest(data: data)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(
+            operation: request,
+            responseDelegate: responseDelegate,
+            completionHandler: completion
+        )
     }
 
-    ///
+    @available(*, deprecated, renamed: "cardList(data:responseDelegate:completion:)")
+    public func сardList(
+        data: InitGetCardListData,
+        responseDelegate: NetworkTransportResponseDelegate?,
+        completionHandler: @escaping (_ result: Result<CardListResponse, Error>) -> Void
+    ) -> Cancellable {
+        cardList(data: data, responseDelegate: responseDelegate, completion: completionHandler)
+>>>>>>> dev
+    }
+
     /// - Parameters:
     ///   - data: `InitAddCardData` информация о клиенте и типе новой карты
-    ///   - completionHandler: результат операции `CardListResponse` в случае удачной регистрации и  `Error` - ошибка.
+    ///   - completion: результат операции `CardListResponse` в случае удачной регистрации и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func сardListAddCardInit(data: InitAddCardData, completionHandler: @escaping (_ result: Result<InitAddCardResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func cardListAddCardInit(
+        data: InitAddCardData,
+        completion: @escaping (_ result: Result<InitAddCardResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = InitAddCardRequest(requestData: data)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completion)
     }
 
-    ///
+    @available(*, deprecated, renamed: "cardListAddCardInit(data:completion:)")
+    public func сardListAddCardInit(
+        data: InitAddCardData,
+        completionHandler: @escaping (_ result: Result<InitAddCardResponse, Error>) -> Void
+    ) -> Cancellable {
+        cardListAddCardInit(data: data, completion: completionHandler)
+>>>>>>> dev
+    }
+
     /// - Parameters:
     ///   - data: `InitAddCardData` информация о клиенте и типе новой карты
-    ///   - completionHandler: результат операции `CardListResponse` в случае удачной регистрации карты и  `Error` - ошибка.
+    ///   - completion: результат операции `CardListResponse` в случае удачной регистрации карты и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func сardListAddCardFinish(data: FinishAddCardData, responseDelegate: NetworkTransportResponseDelegate?, completionHandler: @escaping (_ result: Result<FinishAddCardResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func cardListAddCardFinish(
+        data: FinishAddCardData,
+        responseDelegate: NetworkTransportResponseDelegate? = nil,
+        completion: @escaping (_ result: Result<FinishAddCardResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = FinishAddCardRequest(requestData: data)
         updateCardDataRequestParams(&request.parameters)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(
+            operation: request,
+            responseDelegate: responseDelegate,
+            completionHandler: completion
+        )
     }
 
-    ///
+    @available(*, deprecated, renamed: "cardListAddCardFinish(data:responseDelegate:completion:)")
+    public func сardListAddCardFinish(
+        data: FinishAddCardData,
+        responseDelegate: NetworkTransportResponseDelegate?,
+        completionHandler: @escaping (_ result: Result<FinishAddCardResponse, Error>) -> Void
+    ) -> Cancellable {
+        cardListAddCardFinish(
+            data: data,
+            responseDelegate: responseDelegate,
+            completion: completionHandler
+        )
+>>>>>>> dev
+    }
+
     /// - Parameters:
     ///   - amount: `Double` сумма с копейками
     ///   - requestKey: `String` ключ для привязки карты
-    ///   - completionHandler: результат операции `AddCardStatusResponse` в случае удачной регистрации карты и  `Error` - ошибка.
+    ///   - completion: результат операции `AddCardStatusResponse` в случае удачной регистрации карты и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func chechRandomAmount(_ amount: Double, requestKey: String, responseDelegate: NetworkTransportResponseDelegate?, completionHandler: @escaping (_ result: Result<AddCardStatusResponse, Error>) -> Void) -> Cancellable {
-        let request = CheckRandomAmountRequest(requestData: CheckingRandomAmountData(amount: amount, requestKey: requestKey))
+    @discardableResult
+    public func checkRandomAmount(
+        _ amount: Double,
+        requestKey: String,
+        responseDelegate: NetworkTransportResponseDelegate? = nil,
+        completion: @escaping (_ result: Result<AddCardStatusResponse, Error>) -> Void
+    ) -> Cancellable {
+        let requestData = CheckingRandomAmountData(amount: amount, requestKey: requestKey)
+        let request = CheckRandomAmountRequest(requestData: requestData)
         updateCardDataRequestParams(&request.parameters)
 
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(
+            operation: request,
+            responseDelegate: responseDelegate,
+            completionHandler: completion
+        )
     }
 
-    ///
+    @available(*, deprecated, renamed: "checkRandomAmount(_:requestKey:responseDelegate:completion:)")
+    public func chechRandomAmount(
+        _ amount: Double,
+        requestKey: String,
+        responseDelegate: NetworkTransportResponseDelegate?,
+        completionHandler: @escaping (_ result: Result<AddCardStatusResponse, Error>) -> Void
+    ) -> Cancellable {
+        checkRandomAmount(
+            amount,
+            requestKey: requestKey,
+            responseDelegate: responseDelegate,
+            completion: completionHandler
+        )
+>>>>>>> dev
+    }
+
     /// - Parameters:
-    ///   - completionHandler: результат операции `CardListResponse` в случае удачной регистрации и  `Error` - ошибка.
+    ///   - completion: результат операции `CardListResponse` в случае удачной регистрации и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func сardListDeactivateCard(data: InitDeactivateCardData, completionHandler: @escaping (_ result: Result<FinishAddCardResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func cardListDeactivateCard(
+        data: InitDeactivateCardData,
+        completion: @escaping (_ result: Result<FinishAddCardResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = InitDeactivateCardRequest(requestData: data)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completion)
+    }
+
+    @available(*, deprecated, renamed: "cardListDeactivateCard(data:completion:)")
+    public func сardListDeactivateCard(
+        data: InitDeactivateCardData,
+        completionHandler: @escaping (_ result: Result<FinishAddCardResponse, Error>) -> Void
+    ) -> Cancellable {
+        cardListDeactivateCard(data: data, completion: completionHandler)
+>>>>>>> dev
     }
 
     // MARK: - Система быстрых платежей, оплата по QR-коду
@@ -297,9 +471,20 @@ public final class AcquiringSdk: NSObject {
     ///   - data: `PaymentInvoiceQRCodeData` информация о заказе на оплату
     ///   - completionHandler: результат операции `PaymentInvoiceQRCodeResponse` в случае удачной регистрации и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func paymentInvoiceQRCode(data: PaymentInvoiceQRCodeData, completionHandler: @escaping (_ result: Result<PaymentInvoiceQRCodeResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func paymentInvoiceQRCode(
+        data: PaymentInvoiceQRCodeData,
+        completionHandler: @escaping (_ result: Result<PaymentInvoiceQRCodeResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = PaymentInvoiceQRCodeRequest(data: data)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completionHandler)
+>>>>>>> dev
     }
 
     /// Выставить счет / принять оплату, сгенерировать QR-код для принятия платежей
@@ -308,38 +493,51 @@ public final class AcquiringSdk: NSObject {
     ///   - data: `PaymentInvoiceQRCodeResponseType` информация о заказе на оплату
     ///   - completionHandler: результат операции `PaymentInvoiceQRCodeResponse` в случае удачной регистрации и  `Error` - ошибка.
     /// - Returns: `Cancellable`
-    public func paymentInvoiceQRCodeCollector(data: PaymentInvoiceSBPSourceType, completionHandler: @escaping (_ result: Result<PaymentInvoiceQRCodeCollectorResponse, Error>) -> Void) -> Cancellable {
+    @discardableResult
+    public func paymentInvoiceQRCodeCollector(
+        data: PaymentInvoiceSBPSourceType,
+        completionHandler: @escaping (_ result: Result<PaymentInvoiceQRCodeCollectorResponse, Error>) -> Void
+    ) -> Cancellable {
         let request = PaymentInvoiceQRCodeCollectorRequest(data: data)
+<<<<<<< HEAD
         return sendRequestWithParameters(request, completionHandler: completionHandler)
+=======
+        let commonParameters: JSONObject = createCommonParameters()
+        request.parameters?.merge(commonParameters) { (_, new) -> JSONValue in new }
+
+        return networkTransport.send(operation: request, completionHandler: completionHandler)
+>>>>>>> dev
     }
     
     /// Загрузить список банков, через приложения которых можно совершить оплату СБП
     ///
     /// - Parameters:
     ///   - completion: результат запроса. `SBPBankResponse` в случае успешного запроса и  `Error` - ошибка.
-    
     public func loadSBPBanks(completion: @escaping (Result<SBPBankResponse, Error>) -> Void) {
         let loader = DefaultSBPBankLoader()
         loader.loadBanks(completion: completion)
     }
-    
-    ///
-    
-    public func getTinkoffPayStatus(completion: @escaping (Result<GetTinkoffPayStatusResponse, Error>) -> Void) -> Cancellable {
-        let request = GetTinkoffPayStatusRequest(terminalKey: terminalKey)
-        return networkTransport.send(operation: request) { result in
-            completion(result)
-        }
+
+    @discardableResult
+    public func getTinkoffPayStatus(
+        completion: @escaping (Result<GetTinkoffPayStatusResponse, Error>) -> Void
+    ) -> Cancellable {
+        networkTransport.send(
+            operation: GetTinkoffPayStatusRequest(terminalKey: terminalKey),
+            completionHandler: completion
+        )
     }
-    
-    public func getTinkoffPayLink(paymentId: Int64,
-                                  version: GetTinkoffPayStatusResponse.Status.Version,
-                                  completion: @escaping (Result<GetTinkoffLinkResponse, Error>) -> Void) -> Cancellable {
-        let request = GetTinkoffLinkRequest(paymentId: paymentId,
-                                            version: version)
-        return networkTransport.send(operation: request) { result in
-            completion(result)
-        }
+
+    @discardableResult
+    public func getTinkoffPayLink(
+        paymentId: Int64,
+        version: GetTinkoffPayStatusResponse.Status.Version,
+        completion: @escaping (Result<GetTinkoffLinkResponse, Error>) -> Void
+    ) -> Cancellable {
+        networkTransport.send(
+            operation: GetTinkoffLinkRequest(paymentId: paymentId, version: version),
+            completionHandler: completion
+        )
     }
     
     @discardableResult